Struct gooey_text::Text

#[must_use]
pub struct Text { /* fields omitted */ }

A sequence of [Spans].

Implementations

impl Text

pub fn span<S: Into<String>>(text: S, style: Style) -> Self

Returns a new Text with a single span created from text and style.

pub fn wrap<R: Renderer>(&self, renderer: &R, options: TextWrap) -> PreparedText

Calculates how to render this text and returns the results.

pub fn render_at<R: Renderer>(
&self,
renderer: &R,
location: Point2D<f32, Points>,
wrapping: TextWrap
)

Renders this text at location in renderer. The top-left of the bounding box of the text will be at location.

pub fn render_baseline_at<R: Renderer>(
&self,
scene: &R,
location: Point2D<f32, Points>,
wrapping: TextWrap
)

Renders this text at location in renderer. The baseline of the first line will start at location.

pub fn remove_range(&mut self, range: Range<usize>)

Removes text in range. Empty spans will be removed.

pub fn insert_str(&mut self, offset: usize, value: &str)

Inserts value at offset. Inserts into an existing span.

#[must_use]
pub fn len(&self) -> usize

Returns the total length, in bytes.

#[must_use]
pub fn is_empty(&self) -> bool

Returns true if there are no characters in this text.

pub fn for_each_in_range<F: FnMut(&Span, Range<usize>)>(
&self,
range: Range<usize>,
callback: F
)

Iterates over all spans within range and invokes callback with the span and the applicable range within the span.

pub fn for_each_in_range_mut<F: FnMut(&mut Span, Range<usize>)>(
&mut self,
range: Range<usize>,
callback: F
)

Iterates over all spans within range and invokes callback with the span and the applicable range within the span.